+++ eNMS/static/workflow.js
@@ -706,7 +706,7 @@ function showRestartWorkflowPanel(workflow, service) {
           $("#restart_runtime").append(
             $("<option></option>")
               .attr("value", runtime[0])
-              .text(runtime[0])
+              .text(runtime[1])
           );
         });
         $("#restart_runtime").val(runtimes[runtimes.length - 1]);
@@ -717,7 +717,7 @@ function showRestartWorkflowPanel(workflow, service) {
 }
 
 workflowNamespace.restartWorkflow = function() {
-  fCall(`/run_service/${workflow.id}`, `restart_workflow-form`, function(
+  fCall(`/run_service/${currentPath}`, `restart_workflow-form`, function(
     result
   ) {
     $(`#restart_workflow-${workflow.id}`).remove();
@@ -764,30 +764,36 @@ function displayWorkflowState(result) {
     $("#status").text(`Status: ${result.state.status}`);
     if (result.state.services) {
       $.each(result.state.services, (path, state) => {
-        const id = path.split(">").slice(-1)[0];
-        const color = {
-          true: "#32cd32",
-          false: "#FF6666",
-          skipped: "#D3D3D3",
-          null: "#00CCFF",
-        };
-        if (id in nodes._data && !ends.has(id)) {
-          colorService(id, color[state.success]);
-          const progress = state.progress.device;
-          if (progress.total) {
-            let label = `<b>${nodes.get(id).name}</b>\n`;
-            label += "—————\n";
-            let progressLabel = `Progress - ${progress.success +
-              progress.failure}/${progress.total}`;
-            progressLabel += ` (${progress.success} passed, ${
-              progress.failure
-            } failed)`;
-            label += progressLabel;
-            nodes.update({
-              id: id,
-              label: label,
-            });
+        const id = parseInt(path.split(">").slice(-1)[0]);
+        if (ends.has(id) || !(id in nodes._data)) return;
+        const progress = state.progress.device;
+        colorService(
+          id,
+          progress.skipped == progress.total
+            ? "#D3D3D3"
+            : state.success === true
+            ? "#32cd32"
+            : state.success === false
+            ? "#FF6666"
+            : "#00CCFF"
+        );
+        if (progress.total) {
+          let label = `<b>${nodes.get(id).name}</b>\n`;
+          label += "—————\n";
+          let progressLabel = `Progress - ${progress.success +
+            progress.failure +
+            progress.skipped}/${progress.total}`;
+          let progressInfo = [];
+          if (progress.success) progressInfo.push(`${progress.success} passed`);
+          if (progress.failure) progressInfo.push(`${progress.failure} failed`);
+          if (progress.skipped) {
+            progressInfo.push(`${progress.skipped} skipped`);
           }
+          label += `${progressLabel} (${progressInfo.join(", ")})`;
+          nodes.update({
+            id: id,
+            label: label,
+          });
         }
       });
     }